Extracted resolution process into SpecificationResolver (intermediate step)
